Moved BeaconApi to webapi package
